requirements.txt
# Core application dependencies
streamlit>=1.40.0
altair>=5.3.0
duckdb>=0.9.0
pandas>=2.2.0
numpy>=1.26.4,<2.0

# Environment and configuration
python-dotenv>=1.0.0

# Cloud storage and AWS services
boto3>=1.34.0

# AI/ML providers
anthropic>=0.40.0
google-generativeai>=0.3.0  # For Gemini support

# Authentication (Google OAuth)
requests

# Testing and development
pytest>=7.4.0
pytest-cov>=4.1.0
pytest-mock>=3.12.0
black>=23.0.0
flake8>=6.1.0
isort>=5.12.0
mypy>=1.7.0
types-requests>=2.31.0.10
coverage>=7.3.0

# Optional: Enhanced performance and monitoring
psutil>=5.9.0